requirements.txt
langchain_community
python-dotenv
tiktoken 
langchain-openai
langchainhub
chromadb
beautifulsoup4
youtube-transcript-api
langchain
pytube